refactor: use singular transformation::rotation from OSTk Math rather than plural transformations::rotations
