+++ bindings/python/src/OpenSpaceToolkitAstrodynamicsPy/Flight/Profile/State.cpp
@@ -8,7 +8,7 @@ inline void OpenSpaceToolkitAstrodynamicsPy_Flight_Profile_State(pybind11::modul
 
     using ostk::core::types::Shared;
 
-    using ostk::math::geometry::d3::transformations::rotations::Quaternion;
+    using ostk::math::geometry::d3::transformation::rotation::Quaternion;
     using ostk::math::object::Vector3d;
 
     using ostk::physics::coord::Frame;
