+++ src/OpenSpaceToolkit/Astrodynamics/GuidanceLaw/ConstantThrust.cpp
@@ -15,7 +15,7 @@ namespace astro
 namespace guidancelaw
 {
 
-using ostk::math::geometry::d3::transformations::rotations::Quaternion;
+using ostk::math::geometry::d3::transformation::rotation::Quaternion;
 
 using ostk::astro::trajectory::LocalOrbitalFrameFactory;
 using ostk::astro::trajectory::LocalOrbitalFrameDirection;
